+++ R/dbInterface-ingestPipelines.R
@@ -47,8 +47,8 @@ testForPipeline <- function(pipelineName) {
     error = function(cnd) {
       message(pipelineName, " not found: ", conditionMessage(cnd))  
     })
-  if (exists("res")) {
-    length(res$body[[pipelineName]]) == 1
+  if (exists("res") && is.numeric(res$body[[pipelineName]]$created_date_millis)) {
+    TRUE
   } else {
     FALSE
   }