+++ tests/testthat/test-screening-dbasFileScanning.R
@@ -75,11 +75,11 @@ test_that("False positives are removed", {
 }) 
 
 test_that("Replicate detections are correctly computed", {
-  mergedReport <- getMergedReportDessauBatch()
-  cleanedReport <- cleanReport(getMergedReportDessauBatch(), getRecordsDessauBatch())
-  deletedComps <- setdiff(mergedReport$peakList$comp_name, cleanedReport$peakList$comp_name)
+  reintReport <- getReintegratedReportDessauBatch()
+  cleanedReport <- cleanReport(getReintegratedReportDessauBatch(), getRecordsDessauBatch())
+  deletedComps <- setdiff(reintReport$peakList$comp_name, cleanedReport$peakList$comp_name)
   expect_contains(deletedComps, "Olmesartan-d6")
-  expect_contains(mergedReport$peakList$comp_name, "Methyltriphenylphosphonium")
+  expect_contains(reintReport$peakList$comp_name, "Methyltriphenylphosphonium")
   expect_contains(cleanedReport$peakList$comp_name, "Methyltriphenylphosphonium")
 })