+++ NEWS.md
@@ -3,8 +3,8 @@
 
 ## Major changes (backend)
 
-* Changed table names of screening results tables from `ntsp25.3_dbas*` to `ntsp25.3_feature*`. These tables include
-both DBAS and NTS processing results.
+* Changed table names of screening results tables from `ntsp25.2_dbas*` to `ntsp25.3_feature*`. These tables include
+both DBAS (library screening) and NTS (unknown screening) processing results.
 * New processing method "NTS" to record unknown features (no CSL annotation). Unknown features are added to `feature`
 tables. Unknown features are currently not aligned/grouped, i.e., there is no `name` field. This will be added in a 
 future version.
@@ -23,7 +23,8 @@ statistics, see the "Discover" page at `https://ntsportal.bafg.de` for more info
 * The documentation website is now [online](https://docs.ntsportal.bafg.de/).
 * Two new articles "Document field descriptions for NTSPortal" and "Processing by non-target screening" were added to
 the documentation.
-* Open data licence changed to CC BY 4.0
+* Open data licence changed to CC BY 4.0 for all BfG-measured data.
+
 
 ## Major changes (frontend)
 
@@ -42,6 +43,8 @@ the documentation.
 [documentation.](https://docs.digi.com/resources/documentation/digidocs/90001488-13/reference/r_iso_8601_duration_format.htm)
 * Changed name of `ingest()` to `ingestFeatureRecords()` since this function only has this specific purpose. All other 
 ingesting is done by `appendRecords()`
+* Bug correction: Features found only once in replicate samples (as determined by `replicate_regex`) are removed. The 
+annotation level (1 or 2 [gap-filled]) does not play a role.
 * Vignette documentation moved from vignettes to pkgdown articles, which are available on the documentation website 
 (to reduce installation size).
 * Field names in `msrawfiles` documents have changed to accommodate NTS processing.
@@ -65,7 +68,18 @@ ingesting is done by `appendRecords()`
   - `nts_alias_name`
   - `dbas_alias_name`
   - `nts_alig_mz_tol_units`
-
+* Field names in `spectral_library` documents have changed
+  - `frag_type` to `collision_type`
+* Updated various dependencies to versions:
+  + R = 4.5.2
+  + Python = 3.12.12
+  + ElasticSearch = 9.2.3
+  + elasticsearch (Python) = 9.2.1
+  + CSL = 25.7
+  + ntsworkflow = 0.2.10
+  
+  
+  
 # ntsportal 25.2
 
 ## Major changes (backend and frontend)